Redesigned window and transferred vehicle type pictures
