+++ Projekat/TurnpikeGate/View/ReferentViews/EntranceTicketIssuingForm.cs
@@ -20,12 +20,9 @@ namespace TurnpikeGate.View.ReferentViews
 {
     public partial class EntranceTicketIssuingForm : Form
     {
-        private PictureBox _selectedPicture;
-        private VehicleType _selectedVehicleType;
         private List<System.Threading.Timer> _timers;
         private readonly List<String> _carPlates;
         private readonly ITollStationService _tollStationService;
-        private readonly IPriceListEntryService _priceListEntryService;
         private readonly IPhysicalTollPaymentService _physicalTollPaymentService;
 
         
@@ -33,10 +30,7 @@ namespace TurnpikeGate.View.ReferentViews
         {
 
             InitializeComponent();
-            _selectedPicture = pbCar;
-            _selectedVehicleType = VehicleType.AUTOMOBILE;
             _tollStationService = Globals.Container.Resolve<ITollStationService>();
-            _priceListEntryService = Globals.Container.Resolve<IPriceListEntryService>();
             _physicalTollPaymentService = Globals.Container.Resolve<IPhysicalTollPaymentService>();
             _carPlates = new List<string>();
             
@@ -51,62 +45,22 @@ namespace TurnpikeGate.View.ReferentViews
         {
 
 
-;            PhysicalTollPayment tollPayment = new PhysicalTollPayment(_carPlates[0], DateTime.Now, DateTime.MaxValue,
-                                                                        _selectedVehicleType , _tollStationService.GetAll()[0].ID, ObjectId.Empty, ObjectId.Empty);
+//;            PhysicalTollPayment tollPayment = new PhysicalTollPayment(_carPlates[0], DateTime.Now, DateTime.MaxValue,
+//                                                                        _selectedVehicleType , _tollStationService.GetAll()[0].ID, ObjectId.Empty, ObjectId.Empty);
             
-            _physicalTollPaymentService.Insert(tollPayment);
+//            _physicalTollPaymentService.Insert(tollPayment);
 
-            _carPlates.RemoveAt(0);
+//            _carPlates.RemoveAt(0);
 
-            tbPlates.Text = "";
-            if (_carPlates.Count > 0)
-            {
-                tbPlates.Text = _carPlates[0];
+//            tbPlates.Text = "";
+//            if (_carPlates.Count > 0)
+//            {
+//                tbPlates.Text = _carPlates[0];
 
-            }
+//            }
         }
 
-        private void pbCar_Click(object sender, EventArgs e)
-        {
-            SetSelectedPictureValues(pbCar, VehicleType.AUTOMOBILE);
-        }
-
-        private void SetSelectedPictureValues(PictureBox pictureBox, VehicleType vehicleType)
-        {
-            _selectedPicture.BorderStyle = BorderStyle.None;
-            _selectedPicture.BackColor = Color.Transparent;
-            pictureBox.BorderStyle = BorderStyle.FixedSingle;
-            pictureBox.BackColor = Color.Bisque;
-            
-            _selectedPicture = pictureBox;
-            _selectedVehicleType = vehicleType;
-
-        }
-
-        private void pbBus_Click(object sender, EventArgs e)
-        {
-            SetSelectedPictureValues(pbBus, VehicleType.BUS);
-        }
-
-        private void pbTruck_Click(object sender, EventArgs e)
-        {
-            SetSelectedPictureValues(pbTruck, VehicleType.TRUCK);
-        }
-
-        private void pbMotorcycle_Click(object sender, EventArgs e)
-        {
-            SetSelectedPictureValues(pbMotorcycle, VehicleType.MOTOCYCLE);
-        }
-
-        private void pbMinivan_Click(object sender, EventArgs e)
-        {
-            SetSelectedPictureValues(pbMinivan, VehicleType.MINIVAN);
-        }
-
-        private void pbCarWithTrailer_Click(object sender, EventArgs e)
-        {
-            SetSelectedPictureValues(pbCarWithTrailer, VehicleType.CAR_WITH_TRAILER);
-        }
+        
 
         private void GenerateVehicleThreads()
         {
